requirements.txt
asgiref==3.6.0
astroid==2.5
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
factory-boy==2.12.0
Faker==16.6.1
flake8==3.8.4
idna==2.10
inflection==0.5.1
iniconfig==2.0.0
isort==4.3.21
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.6.1
packaging==23.0
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint-plugin-utils==0.7
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.23.0
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.25.11
wrapt==1.12.1